requirements/base.txt
# Contains requirements for all environments

boto3==1.4.7
botocore==1.7.45
Django==1.11.7
django-markdownx==2.0.21
docutils==0.14
jmespath==0.9.3
Markdown==2.6.9
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
s3transfer==0.1.11
six==1.11.0


